fix(analyses):update files from Leonie Bernet from 24 October 2025

+++ analyses/input/read_cams.py
@@ -25,15 +25,64 @@ parent_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
 if not parent_dir in sys.path:
     sys.path.append(parent_dir)
 from utils import utilities
+#%%
+def cams_unzip(dir_path="/input/ECMWF/CAMS/EAC4_africa"):
+    """
+    Some cams products (e.g. EAC4) are given as zip-files, containing one netcdf per level (e.g. pressure, surface or model level)
+    This function extracts the zip files, merges the containing netcdfs, and saves them as a new netcdf (with the same name as the initial zip-file)
+    """
+
+        
+    file_list = [os.path.join(dir_path, file) for file in os.listdir(dir_path) if file.endswith(".zip")]
+
+    # read data:
+    # loop through all zip-files and extract them
+
+    for file_name in file_list:
+        zip_path = os.path.join(dir_path, file_name)
+        zip_extract = os.path.join(dir_path, os.path.splitext(file_name)[0])
+        new_file_name = zip_extract + ".nc"
+        #skip if the merged nc file does already exist:
+        if os.path.isfile(new_file_name):
+            print(f"File {new_file_name} already exists. Skip.")
+            continue
+        print(f"unzip {file_name}")
+        os.makedirs(zip_extract, exist_ok=True)
+
+        # read and extract the zip file
+        with zipfile.ZipFile(zip_path, "r") as zip_sel:
+            zip_sel.extractall(zip_extract)
+
+        extracted_files_list = os.listdir(zip_extract)
+        #Merge netcdfs if there are several files in the extracted folder
+        cams_temps = []
+        for nc_file in extracted_files_list:
+            cams_temp = xr.open_dataset(os.path.join(zip_extract, nc_file))
+            cams_temps.append(cams_temp)
+        cams_combined = xr.merge(cams_temps)
+        # if the variable "valid_time" is present, rename it to "time"
+        if "valid_time" in cams_combined.coords:
+            cams_combined = cams_combined.rename(
+                {"valid_time": "time"}
+            )        
+        if "model_level" in cams_combined.coords:
+            cams_combined = cams_combined.rename(
+                {"model_level": "level"}
+            )
+        if os.path.isfile(new_file_name):
+            os.remove(new_file_name)
+        cams_combined.to_netcdf(new_file_name)
+        print(f"Merged file saved as {new_file_name}")
+
 
 
 # %%
 def read_cams_inv(
     dir_data,
-    dir_out="..\..\data\cams",
+    dir_out="../../data/cams",
     species="co2",
     yr1=2020,
-    yr2=2023,
+    yr2=2024,
     month1=1,
     month2=12,
     dx=2.5,
@@ -58,8 +107,8 @@ def read_cams_inv(
     # get station coordinates
     lat, lon, alt = utilities.get_station_coords(station)
 
-    dir_path = dir_data + rf"\invGG"
-    # file_list = glob.glob(os.path.join(dir_path + rf'\cams73_latest_{species}_conc_*.nc'))
+    dir_path = dir_data + "/invGG"
+    # file_list = glob.glob(os.path.join(dir_path + rf'/cams73_latest_{species}_conc_*.nc'))
 
     # first, extract all zip files
     file_list = os.listdir(dir_path)
@@ -86,7 +135,7 @@ def read_cams_inv(
     ds_combined = None
     for y in np.arange(yr1, yr2 + 1):
         print(f"read year {y}")
-        data_files = glob.glob(dir_path + rf"\cams73_latest_{species}_conc_*_{y}*.nc")
+        data_files = sorted(glob.glob(dir_path + f"/cams73_latest_{species}_conc_*_{y}*.nc"))
         for file in data_files:
             if y == yr1 and file[-5:-3] < month1_str:
                 # skip months before given first month1 in yr1
@@ -117,7 +166,7 @@ def read_cams_inv(
         ds_combined["CH4"].attrs["units"] = "ppb"
 
     # remove file if already existing
-    fname = rf"{dir_out}\cams_invGG_{species}_{date_start}_{date_stop}_{station}.nc"
+    fname = f"{dir_out}/cams_invGG_{species}_{date_start}_{date_stop}_{station}.nc"
     if os.path.isfile(fname):
         os.remove(fname)
     ds_combined.to_netcdf(fname, mode="w")
@@ -125,7 +174,7 @@ def read_cams_inv(
 
 # %%
 def read_cams_egg4(
-    dir_data, dir_out="..\..\data\cams", yr1=2020, yr2=2020, station="MKN"
+    dir_data, dir_out="../../data/cams", yr1=2020, yr2=2020, station="MKN"
 ):
     """
     Read in the CAMS EGG4 data and save as new netcdf.
@@ -136,7 +185,7 @@ def read_cams_egg4(
 
     for y in np.arange(yr1, yr2 + 1):
         print(f"read year {y}")
-        ds_temp = xr.open_mfdataset(dir_data + rf"\EGG4\cams_egg4_{y}*_{station}.nc")
+        ds_temp = xr.open_mfdataset(dir_data + f"/EGG4/cams_egg4_{y}*_{station}.nc")
 
         # Concatenate along the time dimension to create a single dataset
         if ds_combined is None:
@@ -166,7 +215,7 @@ def read_cams_egg4(
         )  # in ppm or ppb
         ds_combined[var_name] = ds_combined[var_name].assign_attrs(units=unit)
 
-    fname = rf"{dir_out}\cams_egg4_{yr1}_{yr2}_{station}.nc"
+    fname = f"{dir_out}/cams_egg4_{yr1}_{yr2}_{station}.nc"
     if os.path.isfile(fname):
         os.remove(fname)
 
@@ -174,7 +223,7 @@ def read_cams_egg4(
 
 
 # %%
-def read_cams_gfas(dir_data, dir_out="..\..\data\cams", yr1=2020, yr2=2023):
+def read_cams_gfas(dir_data, dir_out="../../data/cams", yr1=2020, yr2=2024,save_combined_netcdf=False):
     """
     Read in all CAMS GFAS biomass burning data and save as new netcdf.
 
@@ -184,29 +233,36 @@ def read_cams_gfas(dir_data, dir_out="..\..\data\cams", yr1=2020, yr2=2023):
 
     for y in np.arange(yr1, yr2 + 1):
         print(f"read year {y}")
-        ds_temp = xr.open_mfdataset(dir_data + rf"\GFAS\cams_gfas_{y}*.nc")
+        ds_temp = xr.open_mfdataset(dir_data + f"/cams_gfas_{y}*.nc")
+        #ds_temp = xr.open_mfdataset(dir_data + f"/GFAS/cams_gfas_{y}*.nc")
 
         # Concatenate along the time dimension to create a single dataset
         if ds_combined is None:
             ds_combined = ds_temp
         else:
-            ds_combined = xr.concat([ds_combined, ds_temp], dim="time")
+            ds_combined = xr.concat([ds_combined, ds_temp], dim="valid_time")
+
+    # Data from new Atmospheric Data Store (ADS) is now sorted by valid_time. Rename to time. 
+    ds_combined = ds_combined.rename({'valid_time': 'time'})
 
     yr1 = ds_combined.time[0].dt.year.values
-    yr2 = ds_combined.time[-1].dt.year.values
+    yr2 = ds_combined.time[-2].dt.year.values # in the new data, the last time step of 31.12. is 01.01. 00:00 of the next year
 
     # rename variables:
     # ds_combined = ds_combined.rename({"co2": "CO2", "ch4": "CH4"})
 
-    fname = rf"{dir_out}\cams\cams_gfas_{yr1}_{yr2}.nc"
-    if os.path.isfile(fname):
-        os.remove(fname)
+    if save_combined_netcdf:
+        fname = f"{dir_out}/cams_gfas_{yr1}_{yr2}.nc"
+        if os.path.isfile(fname):
+            os.remove(fname)
 
-    ds_combined.to_netcdf(fname)
+        ds_combined.to_netcdf(fname)
+
+    return ds_combined
 
 
 # %%
-def read_cams_eac4(dir_data, dir_out="..\..\data\cams", station="MKN"):
+def read_cams_eac4(dir_data, dir_out="../../data/cams", station="MKN"):
     """
     Read in the full cams EAC4 data.
     Two seperate .nc files are saved as a zip file for each month. The files should already contain 9 grids around the MKN station (downloadedf with kadi_get_cams.py)
@@ -215,7 +271,7 @@ def read_cams_eac4(dir_data, dir_out="..\..\data\cams", station="MKN"):
     Save the merged data as new netcdf.
 
     """
-    dir_path = dir_data + rf"\EAC4"
+    dir_path = dir_data + "/EAC4"
     file_list = os.listdir(dir_path)
 
     ds_combined = None
@@ -234,14 +290,29 @@ def read_cams_eac4(dir_data, dir_out="..\..\data\cams", station="MKN"):
                 zip_sel.extractall(zip_extract)
 
             # open the 2 netcdfs in the extracted folder and merge them to one dataset
+            # the file names changed in 2024
+            pfile_name = 'data_plev.nc' if '_2024_' in file_name else 'levtype_pl.nc'
+            sfile_name = 'data_sfc.nc' if '_2024_' in file_name else 'levtype_sfc.nc'
             eac4_pl_temp = xr.open_dataset(
-                rf"{zip_extract}\levtype_pl.nc"
+                f"{zip_extract}/{pfile_name}"  
             )  # pressure level
             eac4_sfc_temp = xr.open_dataset(
-                rf"{zip_extract}\levtype_sfc.nc"
+                f"{zip_extract}/{sfile_name}"  
             )  # single level
 
             eac4_merged_temp = xr.merge([eac4_pl_temp, eac4_sfc_temp])
+
+            if "_2024_" in file_name:
+                eac4_merged_temp = eac4_merged_temp.rename(
+                    {"valid_time": "time", "pressure_level": "level"}
+                )                
+                # Problem: the new 2024 data coordinates are now of type float64, but the old data is of type float32.
+                # This causes problems when merging the datasets, so we convert the coordinates to float32.
+                eac4_merged_temp = eac4_merged_temp.assign_coords({
+                    "latitude": eac4_merged_temp.latitude.astype("float32"),
+                    "longitude": eac4_merged_temp.longitude.astype("float32"),
+                    "level": eac4_merged_temp.level.astype("float32")
+                })
 
             # Concatenate along the time dimension to create a single dataset
             if ds_combined is None:
@@ -271,12 +342,12 @@ def read_cams_eac4(dir_data, dir_out="..\..\data\cams", station="MKN"):
         )  # reassign attributes (not working??)
         ds_combined[var_name] = ds_combined[var_name].assign_attrs(units="ppb")
 
-    fname = rf"{dir_out}\cams_eac4_{yr1}_{yr2}_{station}.nc"
+    fname = f"{dir_out}/cams_eac4_{yr1}_{yr2}_{station}.nc"
     if os.path.isfile(fname):
         os.remove(fname)
     ds_combined.to_netcdf(fname, mode="w")
 
-def read_cams_eac4_aerosols(dir_data, yr1=2020,yr2=2023, dir_out="data\cams", station="MKN"):
+def read_cams_eac4_aerosols(dir_data, yr1=2020,yr2=2023, dir_out="data/cams", station="MKN"):
     """
     Read in the cams EAC4 aerosol mxing ratio data (because that was downloaded seperately, not required if you download all EAC4 data at once).
 
@@ -284,14 +355,25 @@ def read_cams_eac4_aerosols(dir_data, yr1=2020,yr2=2023, dir_out="data\cams", st
     Save the merged data as new netcdf.
 
     """
-    dir_path = dir_data + rf"\EAC4_aerosols"
+    dir_path = dir_data + f"/EAC4_aerosols"
 
     ds_combined = None
     # read data
 
     for y in np.arange(yr1, yr2 + 1):
         print(f"read year {y}")
-        ds_temp = xr.open_mfdataset(dir_path + rf"\cams_eac4_{y}*.nc")
+        ds_temp = xr.open_mfdataset(dir_path + f"/cams_eac4_{y}*.nc")
+        if y==2024:
+            ds_temp = ds_temp.rename(
+                {"valid_time": "time", "pressure_level": "level"}
+            )
+            # Problem: the new 2024 data coordinates are now of type float64, but the old data is of type float32.
+            # This causes problems when merging the datasets, so we convert the coordinates to float32.
+            ds_temp = ds_temp.assign_coords({
+                "latitude": ds_temp.latitude.astype("float32"),
+                "longitude": ds_temp.longitude.astype("float32"),
+                "level": ds_temp.level.astype("float32")
+            })
 
         # Concatenate along the time dimension to create a single dataset
         if ds_combined is None:
@@ -303,7 +385,7 @@ def read_cams_eac4_aerosols(dir_data, yr1=2020,yr2=2023, dir_out="data\cams", st
     yr2 = ds_combined.time[-1].dt.year.values
 
 
-    fname = rf"{dir_out}\cams_eac4_aerosols_{yr1}_{yr2}_{station}.nc"
+    fname = f"{dir_out}/cams_eac4_aerosols_{yr1}_{yr2}_{station}.nc"
     if os.path.isfile(fname):
         os.remove(fname)
     ds_combined.to_netcdf(fname, mode="w")
@@ -311,8 +393,8 @@ def read_cams_eac4_aerosols(dir_data, yr1=2020,yr2=2023, dir_out="data\cams", st
 
 def get_best_cams(
     obs_all,
-    dir_in="..\..\..\Data\CAMS",
-    dir_out="..\..\data\cams",
+    dir_in="../../../Data/CAMS",
+    dir_out="../../data/level3/cams",
     obs_datasets=["CO2", "CH4", "CO", "O3"],
     cams_datasets=[
         "co2_invgg",
@@ -343,21 +425,21 @@ def get_best_cams(
     obs_all_6h = obs_all.resample(time="6h").mean(keep_attrs=True)
 
     # Read all cams datasets (Give the correct filenames!)
-    dir_data_cams = r"../data/cams"
+    dir_data_cams = "../data/level3/cams"
     cams_invgg_co2 = xr.open_dataset(
-        dir_data_cams + r"/cams_invGG_co2_202001_202306_MKN.nc"
+        dir_data_cams + "/cams_invGG_co2_202001_202306_MKN.nc"
     )
     cams_invgg_co2_2 = xr.open_dataset(
-        dir_data_cams + r"/cams_invGG_co2_202307_202309_MKN.nc"
+        dir_data_cams + "/cams_invGG_co2_202307_202412_MKN.nc"
     )
     cams_invgg_ch4 = xr.open_dataset(
-        dir_data_cams + r"/cams_invGG_ch4_202001_202112_MKN.nc"
+        dir_data_cams + "/cams_invGG_ch4_202001_202112_MKN.nc"
     )
     cams_invgg_ch4_2 = xr.open_dataset(
-        dir_data_cams + r"/cams_invGG_ch4_202201_202212_MKN.nc"
+        dir_data_cams + "/cams_invGG_ch4_202201_202412_MKN.nc"
     )
-    cams_eac4 = xr.open_dataset(dir_data_cams + r"/cams_eac4_2003_2023_MKN.nc")
-    cams_egg4 = xr.open_dataset(dir_data_cams + r"/cams_egg4_2003_2020_MKN.nc")
+    cams_eac4 = xr.open_dataset(dir_data_cams + "/cams_eac4_2003_2024_MKN.nc")
+    cams_egg4 = xr.open_dataset(dir_data_cams + "/cams_egg4_2003_2020_MKN.nc")
 
     # for each observational dataset, define the corresponding cams dataset and find the best cams-grid
     datasets = []
@@ -434,7 +516,7 @@ def get_best_cams(
 
     ## merge
     if save_netcdf:
-        fname = rf"{dir_out}\cams_best_grid_MKN.nc"
+        fname = f"{dir_out}/cams_best_grid_MKN.nc"
         if os.path.isfile(fname):
             os.remove(fname)
         cams_best_datasets.to_netcdf(fname, mode="w")
@@ -443,8 +525,8 @@ def get_best_cams(
 
 
 # %%
-def main(dir_data=r"..\..\Data\CAMS", station="MKN"):
-    # when debugging directly in the file: dir_data = r"..\..\Data\CAMS"
+def main(dir_data="../../Data/CAMS", station="MKN"):
+    # when debugging directly in the file: dir_data = "../../Data/CAMS"
     # read_cams_inv(
     #     dir_data,
     #     species="ch4",
